requirements.txt
python-binance==1.0.19
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
plotly==5.20.0
python-dotenv==1.0.1
loguru==0.7.2
matplotlib